requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.111.0
google-auth==2.26.1
APScheduler==3.10.4
python-dotenv==1.0.0